+++ testdata/de/dm/intellij/liferay/language/freemarker/LiferayFtlVariableProviderTest/de/dm/MyService.java
@@ -1,0 +1,8 @@
+package de.dm;
+
+public class MyService {
+
+    public void helloService() {
+
+    }
+}