+++ changelog.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Bugfix: Fixed parsing liferay-look-and-feel.xml when creating that file for the first time
 - Bugfix: Correctly detect locale strings in JSON schema for journal structure files
 - Bugfix: Fixed compatibility with IntelliJ 2018.3 when injecting javascript into freemarker taglibs
+- Bugfix: ServiceLocator lookup in Application Display Templates was not working properly
 - Feature: Code completion for portlet names in JSP taglibs
 - Feature: Support for multiple OSGi service definitions in a @Component annotation
 - Feature: Inspections for service.xml